Minor bug fixes
git-svn-id: https://qooxdoo.svn.sourceforge.net/svnroot/qooxdoo/trunk@15249 61aaf1ca-a10e-0410-b9fb-ac63ec25887f

+++ qooxdoo/framework/source/class/qx/theme/classic/Decoration.js
@@ -309,7 +309,7 @@ qx.Theme.define("qx.theme.classic.Decoration",
 
     "lead-item" :
     {
-      decorator : qx.ui.decoration.Double,
+      decorator : qx.ui.decoration.Single,
 
       style :
       {
